+++ src/shared/config/index.ts
@@ -1,3 +1,2 @@
 export * from './axiosInstance';
 export * from './locale';
-export * from './remotePatterns';
